requirements.txt
neuralhydrology>=1.7.0
numpy>=1.20.0
pandas>=1.3.0
torch>=1.9.0
torchvision>=0.10.0
matplotlib>=3.4.0
seaborn>=0.11.0
scipy>=1.7.0
scikit-learn>=1.0.0
xarray>=0.19.0
netcdf4>=1.5.7
pyyaml>=5.4.0
tensorboard>=2.6.0
jupyter>=1.0.0
plotly>=5.0.0
hydroeval>=0.1.0
tqdm>=4.62.0 